+++ .gitignore
@@ -105,3 +105,5 @@ dist
 
 # WebStorm
 .idea
+
+dist
